refactor: organize commons in more appropriate folders
