+++ src/renderer/src/components/document-views/EmptyDocument.tsx
