+++ src/renderer/src/components/editing/RichTextEditor.tsx
